dhcp lease file path tweaked :memo:

+++ obox_router_configure.py
@@ -106,9 +106,10 @@ def access_point_vanisher():
 
 def connected_device_vanisher():
     for file in os.listdir(dlf):
-        file_path = os.path.join(dlf, file)
-        os.unlink(file_path)
-        print colored("previously connected devices cleared", 'red')        
+        if file.endswith(".leases")
+            file_path = os.path.join(dlf, file)
+            os.unlink(file_path)
+            print colored("previously connected devices cleared", 'red')        
 
 
 if __name__ == "__main__":
